refactor : crew -> challenge (challenge create)
