+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge/create/ChallengeCreate1Fragment.kt
@@ -1,0 +1,12 @@
+package com.side.runwithme.view.challenge.create
+
+import com.example.seobaseview.base.BaseFragment
+import com.side.runwithme.R
+import com.side.runwithme.databinding.FragmentChallengeCreate1Binding
+
+class ChallengeCreate1Fragment : BaseFragment<FragmentChallengeCreate1Binding>(R.layout.fragment_challenge_create1) {
+    override fun init() {
+
+    }
+
+}
