+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge/create/ChallengeCreate2Fragment.kt
@@ -1,0 +1,11 @@
+package com.side.runwithme.view.challenge.create
+
+import com.example.seobaseview.base.BaseFragment
+import com.side.runwithme.R
+import com.side.runwithme.databinding.FragmentChallengeCreate2Binding
+
+class ChallengeCreate2Fragment : BaseFragment<FragmentChallengeCreate2Binding>(R.layout.fragment_challenge_create2) {
+    override fun init() {
+
+    }
+}
