+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge/create/ChallengeCreate3Fragment.kt
@@ -1,0 +1,11 @@
+package com.side.runwithme.view.challenge.create
+
+import com.example.seobaseview.base.BaseFragment
+import com.side.runwithme.R
+import com.side.runwithme.databinding.FragmentChallengeCreate3Binding
+
+class ChallengeCreate3Fragment : BaseFragment<FragmentChallengeCreate3Binding>(R.layout.fragment_challenge_create3) {
+    override fun init() {
+
+    }
+}
