+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge/create/ChallengeCreate4Fragment.kt
@@ -1,0 +1,11 @@
+package com.side.runwithme.view.challenge.create
+
+import com.example.seobaseview.base.BaseFragment
+import com.side.runwithme.R
+import com.side.runwithme.databinding.FragmentChallengeCreate4Binding
+
+class ChallengeCreate4Fragment : BaseFragment<FragmentChallengeCreate4Binding>(R.layout.fragment_challenge_create4) {
+    override fun init() {
+
+    }
+}
