+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge/create/ChallengeCreate5Fragment.kt
@@ -1,0 +1,11 @@
+package com.side.runwithme.view.challenge.create
+
+import com.example.seobaseview.base.BaseFragment
+import com.side.runwithme.R
+import com.side.runwithme.databinding.FragmentChallengeCreate5Binding
+
+class ChallengeCreate5Fragment : BaseFragment<FragmentChallengeCreate5Binding>(R.layout.fragment_challenge_create5) {
+    override fun init() {
+
+    }
+}
