+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge/detail/ChallengeDetailConfirmDialog.kt
@@ -1,10 +1,10 @@
-package com.side.runwithme.view.crew.detail.dialog
+package com.side.runwithme.view.challenge.detail
 
 import com.example.seobaseview.base.BaseDialogFragment
 import com.side.runwithme.R
 import com.side.runwithme.databinding.DialogCrewDetailConfirmBinding
 
-class CrewDetailConfirmDialog : BaseDialogFragment<DialogCrewDetailConfirmBinding>(R.layout.dialog_crew_detail_confirm) {
+class ChallengeDetailConfirmDialog : BaseDialogFragment<DialogCrewDetailConfirmBinding>(R.layout.dialog_crew_detail_confirm) {
     override fun init() {
 
     }
