+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge/detail/ChallengeDetailFragment.kt
@@ -1,0 +1,11 @@
+package com.side.runwithme.view.challenge.detail
+
+import com.example.seobaseview.base.BaseFragment
+import com.side.runwithme.R
+import com.side.runwithme.databinding.FragmentChallengeDetailBinding
+
+class ChallengeDetailFragment : BaseFragment<FragmentChallengeDetailBinding>(R.layout.fragment_challenge_detail) {
+    override fun init() {
+
+    }
+}
