--- RunWithMe/presentation/src/main/java/com/side/runwithme/view/crew/create/CrewCreate1Fragment.kt
@@ -1,11 +1,0 @@
-package com.side.runwithme.view.crew.create
-
-import com.example.seobaseview.base.BaseFragment
-import com.side.runwithme.R
-import com.side.runwithme.databinding.FragmentCrewCreate1Binding
-
-class CrewCreate1Fragment : BaseFragment<FragmentCrewCreate1Binding>(R.layout.fragment_crew_create1) {
-    override fun init() {
-        binding.toolbarMain.setTitle("프로필 ")
-    }
-}
