--- RunWithMe/presentation/src/main/java/com/side/runwithme/view/crew/create/CrewCreate2Fragment.kt
@@ -1,11 +1,0 @@
-package com.side.runwithme.view.crew.create
-
-import com.example.seobaseview.base.BaseFragment
-import com.side.runwithme.R
-import com.side.runwithme.databinding.FragmentCrewCreate2Binding
-
-class CrewCreate2Fragment : BaseFragment<FragmentCrewCreate2Binding>(R.layout.fragment_crew_create2) {
-    override fun init() {
-
-    }
-}
