--- RunWithMe/presentation/src/main/java/com/side/runwithme/view/crew/create/CrewCreate3Fragment.kt
@@ -1,11 +1,0 @@
-package com.side.runwithme.view.crew.create
-
-import com.example.seobaseview.base.BaseFragment
-import com.side.runwithme.R
-import com.side.runwithme.databinding.FragmentCrewCreate3Binding
-
-class CrewCreate3Fragment : BaseFragment<FragmentCrewCreate3Binding>(R.layout.fragment_crew_create3) {
-    override fun init() {
-
-    }
-}
