--- RunWithMe/presentation/src/main/java/com/side/runwithme/view/crew/create/CrewCreate4Fragment.kt
@@ -1,11 +1,0 @@
-package com.side.runwithme.view.crew.create
-
-import com.example.seobaseview.base.BaseFragment
-import com.side.runwithme.R
-import com.side.runwithme.databinding.FragmentCrewCreate4Binding
-
-class CrewCreate4Fragment : BaseFragment<FragmentCrewCreate4Binding>(R.layout.fragment_crew_create4) {
-    override fun init() {
-
-    }
-}
