--- RunWithMe/presentation/src/main/java/com/side/runwithme/view/crew/create/CrewCreate5Fragment.kt
@@ -1,11 +1,0 @@
-package com.side.runwithme.view.crew.create
-
-import com.example.seobaseview.base.BaseFragment
-import com.side.runwithme.R
-import com.side.runwithme.databinding.FragmentCrewCreate5Binding
-
-class CrewCreate5Fragment : BaseFragment<FragmentCrewCreate5Binding>(R.layout.fragment_crew_create5) {
-    override fun init() {
-
-    }
-}
