--- RunWithMe/presentation/src/main/java/com/side/runwithme/view/crew/detail/CrewDetailFragment.kt
@@ -1,11 +1,0 @@
-package com.side.runwithme.view.crew.detail
-
-import com.example.seobaseview.base.BaseFragment
-import com.side.runwithme.R
-import com.side.runwithme.databinding.FragmentCrewDetailBinding
-
-class CrewDetailFragment : BaseFragment<FragmentCrewDetailBinding>(R.layout.fragment_crew_detail) {
-    override fun init() {
-
-    }
-}
