+++ RunWithMe/presentation/src/main/res/layout/dialog_crew_detail_confirm.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.crew.detail.dialog.CrewDetailConfirmDialog">
+        tools:context=".view.challenge.detail.ChallengeDetailConfirmDialog">
 
         <TextView
             android:id="@+id/tv_confirm_text"
