+++ RunWithMe/presentation/src/main/res/layout/fragment_challenge_create5.xml
@@ -8,21 +8,21 @@
         tools:context=".view.crew.create.CrewCreate5Fragment">
 
         <com.side.runwithme.component.CustomToolbar
-            android:id="@+id/toolbar_main"
+            android:id="@+id/toolbar"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             app:back_image="@drawable/baseline_arrow_back_24"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            app:title="챌린지 크루 생성" />
+            app:title="@string/create_challenge" />
 
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            app:layout_constraintBottom_toTopOf="@id/btn_next"
-            app:layout_constraintTop_toBottomOf="@id/toolbar_main"
+            app:layout_constraintBottom_toTopOf="@id/btn_create"
+            app:layout_constraintTop_toBottomOf="@id/toolbar"
             android:layout_marginHorizontal="16dp"
             >
 
@@ -39,24 +39,22 @@
                 android:id="@+id/tv_title"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="챌린지 크루의 \n비밀번호 설정을 해주세요."
+                android:text="챌린지의 \n비밀번호 설정을 해주세요."
                 android:textColor="@color/black"
-                android:textSize="20sp"
+                android:textSize="@dimen/fontsize_h3"
                 android:textStyle="bold"
                 app:layout_constraintTop_toBottomOf="@id/iv_page" />
 
 
             <androidx.constraintlayout.widget.ConstraintLayout
-                android:id="@+id/layout_crew_detail"
+                android:id="@+id/layout_challenge_detail"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintHorizontal_bias="0.0"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/tv_title"
-                app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintVertical_bias="0.2"
-                android:layout_marginBottom="16dp"
                 >
 
 
@@ -64,12 +62,12 @@
                     android:id="@+id/tv_create_crew_password"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:text="비밀번호 설정"
+                    android:text="@string/setting_password"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toTopOf="parent" />
 
                 <RadioGroup
-                    android:id="@+id/radio_group_password"
+                    android:id="@+id/rg_password"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginTop="8dp"
@@ -78,36 +76,74 @@
                     app:layout_constraintTop_toBottomOf="@id/tv_create_crew_password">
 
                     <RadioButton
-                        android:id="@+id/radio_btn_no_pw"
+                        android:id="@+id/rb_no_password"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:buttonTint="@color/main_blue"
-                        android:text="없음"
+                        android:text="@string/no_exist"
                         android:checked="true"
                         android:textColor="@color/black_high_emphasis" />
 
                     <RadioButton
-                        android:id="@+id/radio_btn_pw"
+                        android:id="@+id/rb_password"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:layout_marginStart="72dp"
                         android:buttonTint="@color/main_blue"
-                        android:text="있음"
+                        android:text="@string/exist"
                         android:textColor="@color/black_high_emphasis" />
 
                 </RadioGroup>
 
+
+                <androidx.constraintlayout.widget.ConstraintLayout
+                    android:id="@+id/layout_password_exist"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:visibility="gone"
+                    tools:visibility="visible"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintTop_toBottomOf="@id/rg_password">
+
+                    <TextView
+                        android:id="@+id/tv_password"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginTop="16dp"
+                        android:text="@string/password"
+                        app:layout_constraintStart_toStartOf="parent"
+                        app:layout_constraintTop_toTopOf="parent"/>
+
+                    <androidx.appcompat.widget.AppCompatButton
+                        android:id="@+id/btn_password"
+                        android:layout_width="0dp"
+                        android:layout_height="40dp"
+                        android:layout_gravity="center"
+                        android:layout_marginTop="8dp"
+                        android:background="@drawable/btn_round_white_grey_border"
+                        android:elevation="2dp"
+                        android:maxLength="4"
+                        android:maxLines="1"
+                        android:text="****"
+                        android:textColor="@color/black_low_emphasis"
+                        app:layout_constraintBottom_toBottomOf="parent"
+                        app:layout_constraintEnd_toEndOf="parent"
+                        app:layout_constraintStart_toStartOf="parent"
+                        app:layout_constraintTop_toBottomOf="@id/tv_password"/>
+
+                </androidx.constraintlayout.widget.ConstraintLayout>
 
             </androidx.constraintlayout.widget.ConstraintLayout>
 
         </androidx.constraintlayout.widget.ConstraintLayout>
 
         <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/btn_next"
+            android:id="@+id/btn_create"
             style="@style/StrongButton"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="생성"
+            android:text="@string/create"
             app:layout_constraintBottom_toBottomOf="parent" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
