+++ RunWithMe/presentation/src/main/res/layout/fragment_challenge_detail.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.crew.detail.CrewDetailFragment">
+        tools:context=".view.challenge.detail.ChallengeDetailFragment">
 
         <com.side.runwithme.component.CustomToolbar
             android:id="@+id/toolbar_main"
